+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * Bookings list: Lastname in uppercase
+* Mark as arrived: display number of participants
 
 ### 1.7.2: July 8th, 2021
 * New stat: customers with most no-shows